+++ src/main/java/com/vonage/client/messages/rcs/ImageAlignment.java
@@ -1,0 +1,42 @@
+/*
+ *   Copyright 2025 Vonage
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+package com.vonage.client.messages.rcs;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+
+/**
+ * The alignment of the thumbnail image in a rich card.
+ * This property only applies when sending rich cards with a card_orientation of HORIZONTAL.
+ *
+ * @since 9.9.0
+ */
+public enum ImageAlignment {
+	/**
+	 * Image aligned to the right.
+	 */
+	RIGHT,
+
+	/**
+	 * Image aligned to the left.
+	 */
+	LEFT;
+
+	@JsonValue
+	@Override
+	public String toString() {
+		return name();
+	}
+}