+++ src/main/java/com/vonage/client/messages/rcs/RcsRequest.java
@@ -27,15 +27,26 @@ import com.vonage.client.common.MessageType;
  */
 public abstract class RcsRequest extends MessageRequest {
 	protected Rcs rcs;
+	protected Boolean trustedRecipient;
 
 	protected RcsRequest(Builder<?, ?> builder, MessageType messageType) {
 		super(builder, Channel.RCS, messageType);
 		this.rcs = builder.rcs;
+		this.trustedRecipient = builder.trustedRecipient;
+		int min = 20, max = 259200;
+		if (ttl != null && (ttl < min || ttl > max)) {
+			throw new IllegalArgumentException("TTL must be between "+min+" and "+max+" seconds.");
+		}
 	}
 
 	@JsonProperty("ttl")
 	public Integer getTtl() {
 		return ttl;
+	}
+
+	@JsonProperty("trusted_recipient")
+	public Boolean getTrustedRecipient() {
+		return trustedRecipient;
 	}
 
 	@JsonProperty("rcs")
@@ -46,6 +57,7 @@ public abstract class RcsRequest extends MessageRequest {
 	@SuppressWarnings("unchecked")
 	protected abstract static class Builder<M extends RcsRequest, B extends Builder<? extends M, ? extends B>> extends MessageRequest.Builder<M, B> {
 		protected Rcs rcs;
+		protected Boolean trustedRecipient;
 
 		/**
 		 * (OPTIONAL)
@@ -63,19 +75,6 @@ public abstract class RcsRequest extends MessageRequest {
 
 		/**
 		 * (OPTIONAL)
-		 * Sets the RCS message category.
-		 *
-		 * @param category The RCS category.
-		 * @return This builder.
-		 *
-		 * @since 9.5.0
-		 */
-		public B rcsCategory(String category) {
-			return rcs(new Rcs(category));
-		}
-
-		/**
-		 * (OPTIONAL)
 		 * Indicates if the recipient is trusted.
 		 *
 		 * @param trustedRecipient Whether the recipient is trusted (true or false).
@@ -84,15 +83,20 @@ public abstract class RcsRequest extends MessageRequest {
 		 * @since 9.8.0
 		 */
 		public B trustedRecipient(Boolean trustedRecipient) {
-			if (rcs == null) {
-				rcs = new Rcs();
-			}
-			rcs.trustedRecipient = trustedRecipient;
+			this.trustedRecipient = trustedRecipient;
 			return (B) this;
 		}
 
+		/**
+		 * (OPTIONAL)
+		 * Sets the time-to-live for the RCS message.
+		 *
+		 * @param ttl The duration in seconds the delivery of a message will be attempted,
+		 *            between 300 and 2592000 seconds.
+		 * @return This builder.
+		 */
 		@Override
-		protected B ttl(int ttl) {
+		public B ttl(int ttl) {
 			return super.ttl(ttl);
 		}
 	}